tidied up Utils folder
