+++ src/dpr/components/tag/utils.ts
@@ -1,0 +1,7 @@
+export default {
+  createTagHtml: (text: string, classes?: string) => {
+    return `<p class="govuk-body dpr-tag ${classes}">
+    ${text}
+  </p>`
+  },
+}
